Added controller and all necessary files for shipments. Also added implementation of multiple id search

+++ Services/Dtos/CellStatusDto.cs
@@ -9,7 +9,7 @@ namespace Services.Dtos
 {
     public class CellStatusDto
     {
-        public long Id { get; set; } // Унікальний ID запису
+        public int Id { get; set; } // Унікальний ID запису
 
         public int AlleyId { get; set; }
         public int CellId { get; set; }
